Move `math` module back to top-level
I'm going to need some of the types there from non-kernel modules, and
this is a step in preparation of that.
