+++ src/kernel/mod.rs
@@ -1,6 +1,5 @@
 pub mod approximation;
 pub mod geometry;
-pub mod math;
 pub mod shapes;
 pub mod topology;
 pub mod triangulation;
