+++ src/main.rs
@@ -4,6 +4,7 @@ mod debug;
 mod graphics;
 mod input;
 mod kernel;
+mod math;
 mod mesh;
 mod model;
 mod window;
