+++ src/math/segment.rs
